refactor(keymaps): move LSP related keymaps setup to its own module
